Fix up all cases where we stuff something from a superglobal into a ?int
It's fine (ish) to do this for a string field because PHP will just cope but for
an int or a ?int it's a Fatal error to do this with strict types.

+++ gatherling/Models/Deck.php
@@ -152,7 +152,7 @@ class Deck
                                                                FROM entries INNER JOIN events
                                                                ON entries.event_id = events.id
                                                                WHERE entries.deck = ?', 'd', $this->id);
-            $this->subeventid = Database::single_result_single_param('SELECT id
+            $this->subeventid = (int) Database::single_result_single_param('SELECT id
                                                                    FROM subevents
                                                                    WHERE parent = ?', 's', $this->eventname);
         } else {
